+++ pkg/helm/chart.go
@@ -65,9 +65,12 @@ func GetChaosMeshChartTgzPath(ctx context.Context, version, local string) (strin
 }
 
 func DownloadChaosMeshChartTgz(ctx context.Context, version string) (string, error) {
-	// TODO: use this context
 	url := fmt.Sprintf("%s/chaos-mesh-%s.tgz", ChaosMeshHelmRepo, version)
-	response, err := http.Get(url)
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, url, nil)
+	if err != nil {
+		return "", errors.Wrapf(err, "failed to generate http request for url %s", url)
+	}
+	response, err := http.DefaultClient.Do(req)
 	if err != nil {
 		return "", errors.Wrapf(err, "download helm chart from %s", url)
 	}